+++ src/components/Notification/SimpleNotification/styles.tsx
