+++ src/components/Notification/SimpleNotification/type.ts
@@ -1,4 +1,5 @@
 export interface NotificationItem {
+  projectName?: string;
   notificationId: number;
   createdAt: string;
   content: string;
